+++ src/model/object/object.cpp
@@ -1,5 +1,5 @@
 #include "object.h"
 
-Object::Object(int x, int y) : position{x, y} {}
+Object::Object(double x, double y) : position{x, y} {}
 
 Point Object::getPosition() {return position;}
